Move insert/duplicate/reflection aspects below connection
After all, they are attached to a connection
